+++ KarhooSDK/Service/User/PaymentProviderUpdater.swift
@@ -1,0 +1,9 @@
+//
+// Created by Bartlomiej Sopala on 18/07/2022.
+//
+
+import Foundation
+
+protocol PaymentProviderUpdater {
+    func updatePaymentProvider(user: UserInfo)
+}
